requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
PyGithub==1.59.1
python-dotenv==1.0.0
mangum==0.17.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1
